+++ Guildford/MHSystems/app/src/main/java/com/mh/systems/guildford/ui/activites/BaseActivity.java
@@ -441,4 +441,17 @@ public class BaseActivity extends AppCompatActivity {
         return false;
     }
 
+    /**
+     * To report exception on Roll-bar with user
+     * information.
+     *
+     * @param strNameOfScreen : Name of the Activity/Fragment.
+     * @param strException    : Cause of Exception.
+     */
+    public void reportRollBarException(String strNameOfScreen, String strException) {
+        Rollbar.setPersonData(loadPreferenceValue(ApplicationGlobal.KEY_MEMBERID, "N/A"),
+                loadPreferenceValue(ApplicationGlobal.KEY_USER_LOGINID, "N/A"),
+                null);
+        Rollbar.reportMessage(strNameOfScreen, strException);
+    }
 }