+++ Guildford/MHSystems/app/src/main/java/com/mh/systems/guildford/ui/activites/DashboardActivity.java
@@ -2,12 +2,15 @@ package com.mh.systems.guildford.ui.activites;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.v4.content.ContextCompat;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
@@ -24,22 +27,23 @@ import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
 import com.mh.systems.guildford.R;
+import com.mh.systems.guildford.dataaccess.push.QuickstartPreferences;
+import com.mh.systems.guildford.dataaccess.push.RegistrationIntentService;
 import com.mh.systems.guildford.ui.adapter.RecyclerAdapter.DashboardRecyclerAdapter;
+import com.mh.systems.guildford.utils.SyncMarket;
 import com.mh.systems.guildford.utils.constants.ApplicationGlobal;
+import com.mh.systems.guildford.web.api.WebAPI;
+import com.mh.systems.guildford.web.api.WebServiceMethods;
 import com.mh.systems.guildford.web.models.deletetoken.AJsonParamsDeleteToken;
 import com.mh.systems.guildford.web.models.deletetoken.DeleteTokenAPI;
 import com.mh.systems.guildford.web.models.deletetoken.DeleteTokenResult;
-import com.mh.systems.guildford.web.models.unreadnewscount.AJsonParamsGetUnreadCount;
-import com.mh.systems.guildford.web.models.unreadnewscount.GetUnreadNewsCountAPI;
-import com.mh.systems.guildford.web.models.unreadnewscount.GetUnreadNewsResponse;
 import com.mh.systems.guildford.web.models.featuresflag.AJsonParamsFeaturesFlag;
 import com.mh.systems.guildford.web.models.featuresflag.FeatureFlagsAPI;
 import com.mh.systems.guildford.web.models.featuresflag.FeatureFlagsResponse;
+import com.mh.systems.guildford.web.models.unreadnewscount.AJsonParamsGetUnreadCount;
+import com.mh.systems.guildford.web.models.unreadnewscount.GetUnreadNewsCountAPI;
+import com.mh.systems.guildford.web.models.unreadnewscount.GetUnreadNewsResponse;
 import com.mh.systems.guildford.web.models.weather.WeatherApiResponse;
-import com.mh.systems.guildford.dataaccess.push.QuickstartPreferences;
-import com.mh.systems.guildford.dataaccess.push.RegistrationIntentService;
-import com.mh.systems.guildford.web.api.WebAPI;
-import com.mh.systems.guildford.web.api.WebServiceMethods;
 
 import java.lang.reflect.Type;
 import java.text.SimpleDateFormat;
@@ -150,6 +154,12 @@ public class DashboardActivity extends BaseActivity {
          * automatically cast the corresponding view in your layout.
          */
         ButterKnife.bind(DashboardActivity.this);
+
+        try {
+            checkUpdateVersion();
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+        }
 
         //Initialize adapter.
         dashboardRecyclerAdapter = new DashboardRecyclerAdapter(this,
@@ -374,6 +384,50 @@ public class DashboardActivity extends BaseActivity {
             }
         }
     };
+
+    /**
+     * Check if any update version available on
+     * Google Store.
+     */
+    private void checkUpdateVersion() throws PackageManager.NameNotFoundException {
+        SyncMarket.Initialize(this);
+        final String version = getPackageManager().getPackageInfo(getPackageName(), 0).versionName;
+
+        if(SyncMarket.getMarketVersion() != null && SyncMarket.getMarketVersion().equals(version)){
+            savePreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, version);
+        }
+
+        if (loadPreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, "").equals("")) {
+            savePreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, version);
+        } else {
+            if (!SyncMarket.isVersionEqual(loadPreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, "1.0"))
+                    && SyncMarket.getMarketVersion() != null) {
+                AlertDialog.Builder alertDlg = new AlertDialog.Builder(this);
+                alertDlg.setTitle("Alert");
+                alertDlg.setPositiveButton("Update now", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                        SyncMarket.gotoMarket();
+                    }
+                });
+
+                alertDlg.setNegativeButton("Not now", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        //Store latest version to {@link SharedPreferences} so that never alert for current update.
+                        savePreferenceValue(ApplicationGlobal.KEY_MARKET_VERSION, SyncMarket.getMarketVersion());
+                        dialog.dismiss();
+                    }
+                });
+
+                alertDlg.setMessage(String.format("A New version of " +
+                        getString(R.string.title_golf_club_name) + " (" + SyncMarket.getMarketVersion()
+                        + ") is now available. Please press “UPDATE NOW” to update your current version"));
+                alertDlg.show();
+            }
+        }
+    }
 
 
     /**